Make call_default_handler return a window_event_result
Make call_default_handler return a window_event_result to inform the event system in future, particularly if a window was closed.

+++ common/include/fwd-window.h
@@ -19,7 +19,7 @@ void arch_init();
 namespace dcx {
 
 class window;
-enum class window_event_result : uint8_t;
+enum class window_event_result : unsigned;
 
 template <typename T>
 using window_subfunction = window_event_result (*)(window *menu,const d_event &event, T *userdata);
